+++ cypress/component/basic/styles/css-file/css-file-spec.js
@@ -1,6 +1,6 @@
 /// <reference types="cypress" />
 import React from 'react'
-import { mount } from 'cypress-react-unit-test'
+import { createMount, mount } from 'cypress-react-unit-test'
 
 describe('cssFile', () => {
   it('is loaded', () => {
@@ -63,5 +63,29 @@ describe('cssFile', () => {
       .should(value => {
         expect(parseFloat(value), 'height is < 30px').to.be.lessThan(30)
       })
+  })
+
+  context('Using createMount to simplify global css experience', () => {
+    const mount = createMount({
+      cssFiles: 'cypress/component/basic/styles/css-file/index.css',
+    })
+
+    it('createMount green button', () => {
+      const Component = () => <button className="green">Green button</button>
+      mount(<Component />)
+
+      cy.get('button')
+        .should('have.class', 'green')
+        .and('have.css', 'background-color', 'rgb(0, 255, 0)')
+    })
+
+    it('createMount blue button', () => {
+      const Component = () => <button className="blue">blue button</button>
+      mount(<Component />)
+
+      cy.get('button')
+        .should('have.class', 'blue')
+        .and('have.css', 'background-color', 'rgb(0, 0, 255)')
+    })
   })
 })
