+++ cypress/component/basic/styles/css-file/index.css
@@ -1,3 +1,7 @@
 button.green {
   background-color: #00ff00;
 }
+
+button.blue {
+  background-color: #0000ff;
+}
